single_cnn/requirements.txt
torch~=2.7.1
torchvision~=0.22.1
pillow~=11.2.1
pandas~=2.3.0

numpy~=2.2.6
tqdm~=4.67.1